requirements.txt
Django
requests
djangorestframework
djangorestframework-xml